+++ config/config.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'schema_base_path' => base_path('schemas'),
+];