+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Added support for resolving schemas using a config base path
+
+## Changed
+* Refactored the schema assertions
 
 ## [0.2.0] - 2018-07-17
 ### Changed